backend/requirements.txt
langchain
langgraph
langchain-community
langchain-google-genai
langchain_core
python-dotenv
fastapi
uvicorn
